Migrated plugin to Grails 3.0.x
