+++ .gitignore
@@ -1,3 +1,4 @@
+/grails-cascade-validation.ipr
 /.settings
 /.idea
 /target
