--- application.properties
@@ -1,2 +1,0 @@
-app.grails.version=2.3.11
-app.name=cascade-validation
