--- grails-app/conf/Config.groovy
@@ -1,6 +1,0 @@
-log4j = {
-    error 'org.codehaus.groovy.grails',
-          'org.springframework',
-          'org.hibernate',
-          'net.sf.ehcache.hibernate'
-}
