+++ src/test/groovy/com/cscinfo/platform/constraint/support/ValidateableProperty.groovy
@@ -1,0 +1,14 @@
+package com.cscinfo.platform.constraint.support
+
+import grails.validation.Validateable
+
+/**
+ * @author Eric Kelm
+ */
+class ValidateableProperty implements Validateable {
+    String field
+
+    static constraints = {
+        field blank: false
+    }
+}
