--- test/unit/com/cscinfo/platform/constraint/support/ValidateableParent.groovy
@@ -1,16 +1,0 @@
-package com.cscinfo.platform.constraint.support
-
-import grails.validation.Validateable
-
-/**
- * @author Eric Kelm
- */
-@Validateable
-class ValidateableParent {
-
-    ValidateableProperty property
-
-    static constraints = {
-        property cascade: true
-    }
-}
