--- test/unit/com/cscinfo/platform/constraint/support/ValidateableProperty.groovy
@@ -1,15 +1,0 @@
-package com.cscinfo.platform.constraint.support
-
-import grails.validation.Validateable
-
-/**
- * @author Eric Kelm
- */
-@Validateable
-class ValidateableProperty {
-    String field
-
-    static constraints = {
-        field blank: false
-    }
-}
