--- wrapper/grails-wrapper.properties
@@ -1,1 +1,0 @@
-wrapper.dist.url=http://dist.springframework.org.s3.amazonaws.com/release/GRAILS/
